#176 refactor: 수정 - 패칭된 데이터의 분배 위치 변경
