+++ front/src/components/index.ts
@@ -7,7 +7,7 @@ export { default as Button } from "./Button/Button";
 export { default as Navbar } from "./Navbar/Navbar";
 export { default as filter } from "./filter/filter";
 export { default as ContentCard } from "./ContentCard/ContentCard";
-export { ContentCardList } from "./ContentCardList/ContentCardList";
+export { default as ContentCardList } from "./ContentCardList/ContentCardList";
 export { default as FeedInterface } from "./feed-interface/feed-interface";
 export { default as Comment } from "./comment/comment";
 export { default as ToggleButton } from "./ToggleButton/ToggleButton";
